requirements.txt
python-dotenv==0.21.0
python-telegram-bot==13.14
requests==2.28.1
pathlib2==2.3.7.post1

